[#159] InitialView 삭제
