+++ MC2-OneShot/Presentation/Popup/ShutdownPopupView.swift
@@ -13,13 +13,12 @@ struct ShutdownPopupView: View {
     @Binding var isHelpMessagePresented: Bool
     
     var body: some View {
-        //        VStack(spacing: 0) {
         ZStack(alignment: .center) {
             Rectangle()
                 .frame(width: 364, height: 244)
                 .clipShape(RoundedRectangle(cornerRadius: 15))
                 .foregroundStyle(.shot25).opacity(0.95)
-                .overlay{
+                .overlay {
                     VStack(spacing: 0) {
                         HStack{
                             Image(systemName: "exclamationmark.circle")
@@ -37,17 +36,12 @@ struct ShutdownPopupView: View {
                         
                         Divider()
                         
-                        
                         Text("바로 다음 스탭이 일정 시간동안 완료되지 않아 술자리가 끝난 것 같아 종료했어요 ㅎㅎ!")
-                        
                             .foregroundColor(.shotFF)
                             .pretendard(.regular, 16)
                             .lineSpacing(10)
-                           
                             .padding(.bottom, 56)
                             .padding(.top, 10)
-                        
-                        
                         
                         Spacer()
                         
@@ -60,12 +54,8 @@ struct ShutdownPopupView: View {
                         Spacer()
                     }
                     .padding(.horizontal, 16)
-                    
                 }
-            
-            
         }
-        
         .padding(16)
     }
 }
